Rename main-process/settings to main-process/fragmemoSettings
