+++ main-process/fragmemoSettings/editor.ts
