+++ main-process/fragmemoSettings/window.ts
